Implement field writes closure analysis.
The analysis is implemented in the BodyDecl.fieldWritesClosure attribute.  It
contains a collection of variable accesses.  Each variable accesses (VarAccess)
in this attribute represent a write accesses to a field, that is present in
this body declaration or that can be transitivelly reached from this body
declaration (e.g., over method, constructor, and super-constructor calls).
I.e. the attribute is a transitive closure over all field writes that can be
trigered by executing this body declaration.

IMPORTANT NOTE: This analysis does not consider the implicit
original-ConstructorAccess (as a second statement after the implicit
super-call) in the case of a constructor refinement.  Thus, to get a transitive
closure over all field writes that can be trigered by a given consturctor
refinement, you have to join the output of fieldWritesClosure() for this
constructor refinement with the outputs of fieldWritesClouser() for all other
constructors that might be composed before the given constructor refinement.